+++ javascript/node/selenium-webdriver/builder.js
@@ -128,6 +128,9 @@ class Builder {
 
     /** @private {boolean} */
     this.ignoreEnv_ = false;
+
+    /** @private {http.Agent} */
+    this.agent_ = null;
   }
 
   /**
@@ -185,6 +188,25 @@ class Builder {
    */
   getWebDriverProxy() {
     return this.proxy_;
+  }
+
+  /**
+   * Sets the http agent to use for each request.
+   * If this method is not called, the Builder will use http.globalAgent by default.
+   *
+   * @param {http.Agent} agent The agent to use for each request.
+   * @return {!Builder} A self reference.
+   */
+  usingHttpAgent(agent) {
+    this.agent_ = agent;
+    return this;
+  }
+  
+  /**
+   * @return {http.Agent} The http agent used for each request
+   */
+  getHttpAgent() {
+    return this.agent_;
   }
 
   /**
@@ -454,7 +476,7 @@ class Builder {
     }
 
     if (url) {
-      var executor = executors.createExecutor(url, this.proxy_);
+      var executor = executors.createExecutor(url, this.agent_, this.proxy_);
       return WebDriver.createSession(executor, capabilities, this.flow_);
     }